Narrow mdlFile 0033a220 pointer parameter

+++ src/Event/mt_evtLoadSave.c
@@ -24,7 +24,7 @@ extern u32 LAB_00392bb0;
 
 
 /* Region 0x390000-0x3CFFFF recovered prototypes */
-void FUN_0038dad0(u32 param_1,long param_2,u64 param_3);
+void FUN_0038dad0(u32 param_1,int param_2,int param_3);
 void FUN_0038e660(int param_1,int param_2,int param_3);
 u32 FUN_0038e860(u32 param_1,u32 param_2);
 void FUN_0038f0f0(u32 param_1);
@@ -72,7 +72,7 @@ u32 FUN_003912f0(u32 param_1,u32 param_2);
 // FUN_0038DAD0 NONMATCHING
 
 
-void FUN_0038dad0(u32 param_1,long param_2,u64 param_3)
+void FUN_0038dad0(u32 param_1,int param_2,int param_3)
 
 
 
@@ -2291,7 +2291,7 @@ void FUN_0038fa10(int param_1,u64 param_2)
 // FUN_0038FFB0 NONMATCHING
 
 
-void FUN_0038ffb0(int param_1,u64 param_2)
+void FUN_0038ffb0(int param_1,int param_2)
